Menambah Search Filter dari Kekuatan, Kecepatan & Pagination, & Add Role untuk Char

+++ database/migrations/2025_06_08_144056_create_characters_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('rank');
-            $table->unsignedTinyInteger('power');
-            $table->unsignedTinyInteger('speed');
-            $table->unsignedTinyInteger('jump');
-            $table->unsignedTinyInteger('deff');
+            $table->Integer('power');
+            $table->Integer('speed');
+            $table->Integer('jump');
+            $table->Integer('deff');
             $table->timestamps();
         });
     }
